[XY] Order by metric/custom metric doesn't work well for split chart
